Chapter 09: Test Doubles
Sometimes it is just plain hard to test the system under test (SUT)
because it depends on other components that cannot be used in the test
environment. This could be because they aren't available, they will not
return the results needed for the test or because executing them would
have undesirable side effects. In other cases, our test strategy
requires us to have more control or visibility of the internal behavior
of the SUT.

When we are writing a test in which we cannot (or chose not to) use a
real depended-on component (DOC), we can replace it with a Test Double.
The Test Double doesn't have to behave exactly like the real DOC; it
merely has to provide the same API as the real one so that the SUT
thinks it is the real one!